pyproject.toml
[project]
name = "datamatch"
version = "0.1.0"
description = "Python package for DUNS number matching using D&B Identity Resolution API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.2.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "numpy>=1.24.0",
    "SQLAlchemy>=2.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "tqdm>=4.66.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.scripts]
duns-match = "dunsMatchAPI.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["dunsMatchAPI"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
